requirements.txt
PyYAML>=6.0
alpaca-py>=0.43.0
requests>=2.32.0
pandas>=2.0.0

# Development dependencies
pytest>=7.0
